fix: 編集機能を再び修正
